Ajustes no readme e nos testes.
